backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.105.0
psycopg2-binary==2.9.9
python-dotenv==1.0.7
pymongo==4.6.1
anthropic==0.8.1
